requirements.txt
streamlit
openai
python-dotenv
notion-client

